refactor: `run` -> `sync` & `sync` -> `load`
